Move providers tests to corresponding package and some refactors
