+++ owncloudData/src/test/java/com/owncloud/android/data/providers/implementation/OCSharedPreferencesProviderTest.kt
@@ -1,9 +1,8 @@
-package com.owncloud.android.data.preferences.datasource.implementation
+package com.owncloud.android.data.providers.implementation
 
 import android.content.Context
 import android.content.SharedPreferences
 import android.preference.PreferenceManager
-import com.owncloud.android.data.providers.implementation.OCSharedPreferencesProvider
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.mockkStatic
